backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
yfinance==0.2.51
pandas==2.2.3
python-dateutil==2.9.0
gunicorn==23.0.0
